rename `user-error.ts` -> `error-response.ts`
